SQS move to client package
