+++ client/sqs/message.go
