+++ client/sqs/message_test.go
