+++ client/sqs/publisher.go
