+++ client/sqs/publisher_integration_test.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sqs"
 	"github.com/aws/aws-sdk-go/service/sqs/sqsiface"
@@ -88,12 +87,10 @@ func Test_Publish_Message(t *testing.T) {
 }
 
 func createAPI(t *testing.T) sqsiface.SQSAPI {
-
 	sess, err := session.NewSession(
 		aws.NewConfig().
 			WithEndpoint(testSqsEndpoint).
-			WithRegion(testSqsRegion).
-			WithCredentials(credentials.NewStaticCredentials("test", "test", "")),
+			WithRegion(testSqsRegion),
 	)
 	require.NoError(t, err)
