+++ client/sqs/publisher_test.go
